Add files via upload
This is just the paper for the dataset cause it's too big. There is a link in the paper for the dataset. The number of participants is not 100 but it's 65 (33 males 32 females) in case we don't find higher. There is also multi class classification with 3 (no anxiety, low anxiety, high anxiety).